Refactoring grahp persistence
