+++ packages/nexusgraph-redux/index.ts
@@ -1,6 +1,6 @@
 // Copyright 2023 Paion Data. All rights reserved.
 export type { GlobalState } from "./src/globalState";
-export * from "./src/nlp/nlpDuck";
+export * from "./src/graph/nlpDuck";
 export * from "./src/note-list/noteListDuck";
 export * from "./src/note/noteDuck";
 export * from "./src/oAuth/oAuthDuck";
