+++ packages/nexusgraph-redux/src/graph/nlpDuck.ts
